Reload source comments for all users currently viewing that source
This also requires a modification to the CustomMessageHandler, which
now takes two constructor arguments:

```
new CustomMessageHandler(dispatch, getState)
```

The change allows the message handler to consider the current frontend
state when deciding on how to handle an incoming message.

E.g., when a comment is added, we send a message from the backend,
saying "Please refresh source XYZ".  The message handler, having
access to the app state, checks whether the user is viewing XYZ before
issuing a `fetchSource` action.