Restyle entire profile page using Material UI
This also changes the token form to use react-hook-form

Token fields in the form now are organized as an array, making it
easier to translate the submitted values into a list of ACLs.  I
therefore also modified the createToken endpoint to accept a list of
ACLs, instead of having to parse `acls_id` fields.