Update Javascript `API(...)` to return additional error state
An error can occur for two reasons: the backend explicitly errored, or
the backend gave up and raised an exception.  In the first case, we
typically have an error message, whereas in the latter we just know
something went wrong.

This change allows for these scenarios to be treated
differently (although it remains easy to treat them the same).