Group management reskin (#439)
* Update Main.jsx.template so compiled version passes linter

* Add linting for react hooks

* Update Javascript `API(...)` to return additional error state

An error can occur for two reasons: the backend explicitly errored, or
the backend gave up and raised an exception.  In the first case, we
typically have an error message, whereas in the latter we just know
something went wrong.

This change allows for these scenarios to be treated
differently (although it remains easy to treat them the same).

* Correctly render class tag of ThumbnailList

The class statement sometimes evaluated to "false", which is invalid.

* Distinguish between different source loading stages

If there is an error provided by the backend while loading the source,
display it.

* Start restyling groups pages

- When an invalid source was loaded, an attempt was made to update
sourceview table
- Ported group management interface, partially, to Material UI
- Refactored logic for checking which users are administrators of a group

* Upgrade eslint-plugin-react-hooks to 4.0.0

* Add missing useEffect dependencies

Strictly speaking, `useDispatch` will never change, so doesn't have to
be in there, but the linter isn't smart enough to know about it.