Upgrade to Bokeh 2.
This is the closest I could get to making it run.

A remaining issue is that our custom models are not registered.
Recommended approach: translate models to JavaScript, and have it
registered through register_models from `@bokehjs/base`.

This PR already translates our models to TypeScript.  But, for the
suggested approach we would either need to depend on TypeScript or,
more likely, transcribe the model to JavaScript.

For a TypeScript model extension example, see:

https://github.com/holoviz/panel/blob/master/panel/models/location.ts