Remove session before returning in PS1 callback
This addresses a database session leak.
Each handler (request) gets its own session. The handler is responsible
for closing the session (and, by implication, associated database connection)
when it is done.  This happens in `on_finish`.
When a handler launches a task using `add_callback`, that task could
a) access the database and b) continue operating after the handler's
`on_finish` method had been called.  This would result in the session's
database connection remaining active, and not being returned to the
SQLAlchemy thread pool.
Such connection leaks result in the following error:
  QueuePool limit of size 5 overflow 10 reached, connection timed out,
  timeout 30 (Background on this error at: http://sqlalche.me/e/13/3o7r)