Use Flow directly when firing off task to add ps1 thumbnail
The request may no longer be around, and we should not keep it alive
just for the sake of the task.
We should connect directly to the websocket bus to send messages to
the relevant frontends.
Errors cannot be propagated back to the handler, so that section is
removed and is now logged instead.