GA: fix caching of pip and npm (#2155)
* GA: fix caching of pip and npm

Previously, the npm cache key was generated from **/lockfiles, which we don't use.

This cache key is also not a valid way of invalidating the pip cache.  There are now two caches:
one for NPM (keyed with package.json) and one for pip (keyed with requirements.txt).

* Check out repository before hashing its contents

Otherwise, the hash keys will always be empty.