test_rate_limiting: expect rate limiting earlier
Even with burst mode enabled, on a fast machine we will see rate
limiting being applied soon after the burst buckets are filled.  Lower
the number of successful requests to accommodate this on fast
machines.

On slower machines, such as CI, we should still expect to hit the 16 mark.

Refer to https://www.nginx.com/blog/rate-limiting-nginx/#bursts for
more detail on how rate limiting is applied in nginx.