Upgrade requests and patch its `requests.compat.json` module to handle nan/inf (#2251)
* Pin baselayer

* Allow requests v2.26

* Pin baselayer

* Monkey-patch `simplejson.dumps` to ignore the `allow_nan` keyword

Requests recently started insisting that JSON not contain inf and nan
values in strict compliance with the JSON standard. This makes our
life difficult since scientific data inevitably contains those.

Here, we monkey patch the `requests.compat.json` module (typically,
that translates to `simplejson`) to ignore the `allow_nan` keyword to
`dumps` so that those values pass through as with earlier versions of
requests.

The relevant requests PR is https://github.com/psf/requests/issues/5767

* test_upload_spectroscopy: be less specific about uploaded file MIME type

Different machines will attach different MIME types to files with `.ascii` extensions.
Instead of waiting for the mime type to appear on upload, wait for the filename to appear.
This makes it possible to run this test on machines with different MIME mappings.

* test_rate_limiting: expect rate limiting earlier

Even with burst mode enabled, on a fast machine we will see rate
limiting being applied soon after the burst buckets are filled.  Lower
the number of successful requests to accommodate this on fast
machines.

On slower machines, such as CI, we should still expect to hit the 16 mark.

Refer to https://www.nginx.com/blog/rate-limiting-nginx/#bursts for
more detail on how rate limiting is applied in nginx.

* Also patch requests (to ignore `allow_nan`) in data loader