Correctly configure allocation -> requests relationship
This patch addresses a mis-configuration in the allocation ->
followup requests relationship whereby an error would occur if
attempting to delete an allocation via `DBSession.delete`, which
requires `passive_deletes=True` to be specified on the relationship
pointing to the child class. An API test is added to ensure this now
works as expected (this test will fail without the relationship change).