Add LocalizationTile table to determine which objects are inside localization (#2575)
This PR adds a `LocalizationTile` table (with associated healpix id and probability densities) to determine which objects are inside skymaps. It also adds healpix to the Obj table. 

This PR exposes [GeoJSON](https://geojson.org/) for use in visualization tooling. The form of the `["data"]["geojson"]` packet returned when on a GET to `galaxy_catalog` is a list of `dict`:

```
[
 {'geometry': {'coordinates': [0.60321, -54.76653], 'type': 'Point'},
  'properties': {'name': '2MASX J00022478-5445592'},
  'type': 'Feature'},
 {'geometry': {'coordinates': [0.948375, -54.574083], 'type': 'Point'},
  'properties': {'name': 'IRAS F00012-5451 ID'},
  'type': 'Feature'}
]
```

Co-authored-by: Joshua Bloom <profjsb@gmail.com>
Co-authored-by: Stefan van der Walt <sjvdwalt@gmail.com>