Add custom bandpasses (#2581)
This PR adds two custom sncosmo.Bandpass filters to SkyPortal. It adds the filters via data in the config and loads the bandpasses ENUM with the new filter names. It also adds to the sncosmo registry so that the central wavelength is available to the plot.py color picker. This requires a migration.

This PR supercedes #2549, which contemplated using API endpoints to add bandpasses to the DB. Since adding filters will not be that common, and requires upgrading the enums, adding new filters via the config seems to be the fastest path (and least risky). This PR will be needed for the ATLAS PR #2498.

Co-authored by: jadalilleboe
Co-authored-by: Stefan van der Walt <sjvdwalt@gmail.com>