Fixed incorrect conversion from local to UTC time in the frontend.
A form has a default value, and when submitting it without modifying a start and end date, all is well, and the date ends in "+00:00" as expected, which we delete. But right now, it seems that modifying the date manually or via the calendar that pops up, modifies the format of the date which then ends int'.000Z' that is not removed.

This takes care of that problem by also removing '.000Z'.