Telescope Handler - Performance Improvements and Bugfix (#3101)
This PR does 2 things: (i) reduce the number of calls to the functions to calculate the next morning astronomical and next evening astronomical of a telescope (used to just call the same function multiple times in a row, so here instead we call it once and save its result in a variable, on which we perform operations) and (ii) add a try-except statement if the value returned by those 2 functions is not what we expect, or if nothing can be returned. If that happens, we do the same as what we do for a non-fixed telescope: we set the next morning and evening astronomical to false and continue.