GCN Summary (#3156)
This PR adds a "Gcn Summary" handler, which generates text describing the activity related to an event, which can be used to produce a circular, or just gather data in general. The summary is composed of tables for the sources (and associated photometry), galaxies and observations that are within the event's localization.